Change images directory to Cloudinary since free heroku will automatically delete photos, and make an author unable to modify others post.

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,7 @@ use App\Category, App\Photo, App\User, App\Post;
 use App\Http\Requests\PostsRequest;
 use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Auth;
+use JD\Cloudder\Facades\Cloudder;
 
 class PostsController extends Controller
 {
@@ -43,16 +44,19 @@ class PostsController extends Controller
         $input = $request->all();
 
         if ($file = $request->file('photo')){
-            $name = $file->getClientOriginalName();
+            /* Trim file extension. */
+            $extension = ".".$file->getClientOriginalExtension();
+            $name = substr($file->getClientOriginalName(), 0, -strlen($extension));
             /* Save file name into database first. */
             $photo = Photo::create(['file' => $name]);
             /* Slug id with file name to avoid photos with same file name
                unlinked at the delete moment. */
-            $name = strval($photo->id)."_".substr($photo->file, 8);
+            $name = strval($photo->id)."_".$photo->file;
             /* Update the slugged file name into database. */
             $photo->update(['file' => $name]);
             /* Copy file into /public/images. */
-            $file->move($photo->directory, $name);
+            // $file->move($photo->directory, $name);
+            Cloudder::upload($file, $name, ['folder' => $photo->post_directory]);
             /* Save photo id into posts table. */
             $input['photo_id'] = $photo->id;
         }
@@ -81,8 +85,13 @@ class PostsController extends Controller
     public function edit($slug)
     {
         $post = Post::findBySlugOrFail($slug);
-        $categories = Category::pluck('name', 'id')->all();
-        return view('posts.edit', compact('post', 'categories'));
+        if (Auth::user()->id == $post->user->id){
+            $categories = Category::pluck('name', 'id')->all();
+            return view('posts.edit', compact('post', 'categories'));
+        }
+        else{
+            return redirect('/home/posts/'.$post->slug);
+        }
     }
 
     /**
@@ -95,43 +104,52 @@ class PostsController extends Controller
     public function update(PostsRequest $request, $id)
     {
         // return $request->all();
-        $input = $request->all();
         $post = Post::findOrFail($id);
 
-        if ($file = $request->file('photo')){
-            $name = $file->getClientOriginalName();
-            if (count($post->photo) > 0){
-                /* Remove the file in /public/images first. */
-                unlink(public_path().$post->photo->file);
-                /* Slug id with file name to avoid photos with same file name
-                   unlinked at the delete moment. */
-                $name = strval($post->photo_id)."_".$name;
-                /* Update the slugged file name into database. */
-                Photo::findOrFail($post->photo_id)->update(['file' => $name]);
-                /* Save the file in /public/images. */
-                $file->move($post->photo->directory, $name);
+        if (Auth::user()->id == $post->user->id){
+            $input = $request->all();
+
+            if ($file = $request->file('photo')){
+                /* Trim file extension. */
+                $extension = ".".$file->getClientOriginalExtension();
+                $name = substr($file->getClientOriginalName(), 0, -strlen($extension));
+
+                if (count($post->photo) > 0){
+                    /* Remove the file in /public/images first. */
+                    // unlink(public_path().$post->photo->file);
+                    Cloudder::destroy($post->photo->file, ['folder' => $post->photo->post_directory, 'invalidate' => true]);
+                    /* Slug id with file name to avoid photos with same file name
+                    unlinked at the delete moment. */
+                    $name = strval($post->photo_id)."_".$name;
+                    /* Update the slugged file name into database. */
+                    Photo::findOrFail($post->photo_id)->update(['file' => $name]);
+                    /* Save the file in /public/images. */
+                    // $file->move($post->photo->directory, $name);
+                    Cloudder::upload($file, $name, ['folder' => $post->photo->post_directory]);
+                }
+                else{
+                    /* Save file name into database first. */
+                    $photo = Photo::create(['file' => $name]);
+                    /* Slug id with file name to avoid photos with same file name
+                    unlinked at the delete moment. */
+                    $name = strval($photo->id)."_".substr($photo->file, 8);
+                    /* Update the slugged file name into database. */
+                    $photo->update(['file' => $name]);
+                    /* Save the file in /public/images. */
+                    // $file->move($photo->directory, $name);
+                    Cloudder::upload($file, $name, ['folder' => $photo->post_directory]);
+                    /* Save photo id into posts table. */
+                    $input['photo_id'] = $photo->id;
+                }
             }
-            else{
-                /* Save file name into database first. */
-                $photo = Photo::create(['file' => $name]);
-                /* Slug id with file name to avoid photos with same file name
-                   unlinked at the delete moment. */
-                $name = strval($photo->id)."_".substr($photo->file, 8);
-                /* Update the slugged file name into database. */
-                $photo->update(['file' => $name]);
-                /* Save the file in /public/images. */
-                $file->move($photo->directory, $name);
-                /* Save photo id into posts table. */
-                $input['photo_id'] = $photo->id;
+            /* Update the category id of the post if changed. */
+            if ($input['category_id'] != $post->category_id){
+                Session::flash('change_category', 'Your post "'.$post->title
+                    .'" in Flow "'.$post->category->name.'" has been changed to Flow "'
+                    .Category::findOrFail($input['category_id'])->name.'" now!');
             }
+            $post->update($input);
         }
-        /* Update the category id of the post if changed. */
-        if ($input['category_id'] != $post->category_id){
-            Session::flash('change_category', 'Your post "'.$post->title
-                .'" in Flow "'.$post->category->name.'" has been changed to Flow "'
-                .Category::findOrFail($input['category_id'])->name.'" now!');
-        }
-        $post->update($input);
         return redirect('/home/posts/'.$post->slug);
     }
 
@@ -144,15 +162,21 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
-        $post_category = $post->category;
-        Session::flash('delete_your_post', 'Your post "'.$post->title.'" in Flow "'.$post_category->name.'" has been deleted successfully!');
-        if (count($post->photo) > 0){
-            /* Delete post photo from images storage path. */
-            unlink(public_path().$post->photo->file);
-            /* Delete post photo record from database. */
-            $post->photo->delete();
+        if (Auth::user()->id == $post->user->id){
+            $post_category = $post->category;
+            Session::flash('delete_your_post', 'Your post "'.$post->title.'" in Flow "'.$post_category->name.'" has been deleted successfully!');
+            if (count($post->photo) > 0){
+                /* Delete post photo from images storage path. */
+                // unlink(public_path().$post->photo->file);
+                Cloudder::destroy($post->photo->file, ['folder' => $post->photo->post_directory, 'invalidate' => true]);
+                /* Delete post photo record from database. */
+                $post->photo->delete();
+            }
+            $post->delete();
+            return redirect('/home/'.$post_category->slug);
         }
-        $post->delete();
-        return redirect('/home/'.$post_category->slug);
+        else{
+            return redirect('/home/posts/'.$post->slug);
+        }
     }
 }
